Changed list types, added Command struct, and modified Select functions
Again changed list types to Generic Lists. Created a custom Struct for
Commands which will make it easy to modify them for unit's use. Also,
modified selected functions to include a peek function which just sets
the planeRenderer to select value. Selected is meant to be overriden if
there needs to be a custom select event.